Made key_counts be an int64_t.

+++ src/btree/get_distribution.hpp
@@ -9,6 +9,6 @@
 
 class superblock_t;
 
-void get_btree_key_distribution(btree_slice_t *slice, transaction_t *txn, superblock_t *superblock, int depth_limit, int *key_count_out, std::vector<store_key_t> *keys_out);
+void get_btree_key_distribution(btree_slice_t *slice, transaction_t *txn, superblock_t *superblock, int depth_limit, int64_t *key_count_out, std::vector<store_key_t> *keys_out);
 
 #endif /* BTREE_GET_DISTRIBUTION_HPP_ */
