+++ src/memcached/queries.hpp
@@ -97,9 +97,8 @@ struct distribution_result_t  {
     //Then k1 == left_key
     //and key_counts[ki] = the number of keys in [ki, ki+1) if i < n
     //key_counts[kn] = the number of keys in [kn, right_key)
-    // TODO: Just make this use an int64_t.
     hash_region_t<key_range_t> region;
-    std::map<store_key_t, int> key_counts;
+    std::map<store_key_t, int64_t> key_counts;
 };
 
 /* `gets` */
